refactor: Moved to storage
