+++ score-client/src/main/java/bio/overture/score/client/command/RepositoryAccessCommand.java
@@ -19,7 +19,7 @@ package bio.overture.score.client.command;
 
 import java.io.IOException;
 
-import bio.overture.score.client.transport.StorageService;
+import bio.overture.score.client.storage.StorageService;
 import org.springframework.beans.factory.annotation.Autowired;
 
 public abstract class RepositoryAccessCommand extends AbstractClientCommand {
