+++ score-client/src/main/java/bio/overture/score/client/storage/Gen3StorageService.java
@@ -1,4 +1,4 @@
-package bio.overture.score.client.transport;
+package bio.overture.score.client.storage;
 
 import bio.overture.score.client.config.ClientProperties;
 import bio.overture.score.client.download.DownloadStateStore;
