+++ score-client/src/main/java/bio/overture/score/client/transport/Transport.java
@@ -18,6 +18,7 @@
 package bio.overture.score.client.transport;
 
 import bio.overture.score.client.progress.Progress;
+import bio.overture.score.client.storage.StorageServiceInterface;
 import bio.overture.score.core.model.Part;
 
 import java.io.File;
