+++ score-client/src/main/java/bio/overture/score/client/upload/s3/S3UploadService.java
@@ -20,7 +20,7 @@ package bio.overture.score.client.upload.s3;
 import bio.overture.score.client.exception.NotResumableException;
 import bio.overture.score.client.exception.NotRetryableException;
 import bio.overture.score.client.progress.Progress;
-import bio.overture.score.client.transport.StorageService;
+import bio.overture.score.client.storage.StorageService;
 import bio.overture.score.client.transport.Transport;
 import lombok.SneakyThrows;
 import lombok.val;
